feat(per): Per solidity sdk (#1325)
* feat(per): add solidity sdk

* refactor(evm): refactor abi generator script

* feat(evm): add abi_generator package
