+++ express_relay/sdk/js/package.json
@@ -2,10 +2,8 @@
   "name": "@pythnetwork/express-relay-evm-js",
   "version": "0.1.1",
   "description": "Utilities for interacting with the express relay protocol",
-  "homepage": "https://pyth.network",
-  "author": {
-    "name": "Pyth Data Association"
-  },
+  "homepage": "https://github.com/pyth-network/pyth-crosschain/tree/main/express_relay/sdk/js",
+  "author": "Douro Labs",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "files": [
