+++ express_relay/sdk/solidity/README.md
@@ -1,0 +1,27 @@
+# Pyth Express Relay Solidity SDK
+
+## Install
+
+### Truffle/Hardhat
+
+If you are using Truffle or Hardhat, simply install the NPM package:
+
+```bash
+npm install @pythnetwork/per-sdk-solidity
+```
+
+### Foundry
+
+If you are using Foundry, you will need to create an NPM project if you don't already have one.
+From the root directory of your project, run:
+
+```bash
+npm init -y
+npm install @pythnetwork/per-sdk-solidity
+```
+
+Then add the following line to your `remappings.txt` file:
+
+```text
+@pythnetwork/per-sdk-solidity/=node_modules/@pythnetwork/per-sdk-solidity
+```
