+++ express_relay/sdk/solidity/abis/IPERMulticall.json
@@ -1,0 +1,26 @@
+[
+  {
+    "inputs": [
+      {
+        "internalType": "address",
+        "name": "protocol",
+        "type": "address"
+      },
+      {
+        "internalType": "bytes",
+        "name": "permissionKey",
+        "type": "bytes"
+      }
+    ],
+    "name": "isPermissioned",
+    "outputs": [
+      {
+        "internalType": "bool",
+        "name": "permissioned",
+        "type": "bool"
+      }
+    ],
+    "stateMutability": "view",
+    "type": "function"
+  }
+]
