+++ express_relay/sdk/solidity/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@pythnetwork/per-sdk-solidity",
+  "version": "0.1.0",
+  "description": "Solidity SDK for interacting with Pyth express relay contracts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/pyth-network/pyth-crosschain",
+    "directory": "express_relay/sdk/solidity"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "format": "npx prettier --write .",
+    "generate-abi": "npx generate-abis IPERMulticall IPERFeeReceiver",
+    "check-abi": "git diff --exit-code abis"
+  },
+  "keywords": [
+    "pyth",
+    "oracle",
+    "relay"
+  ],
+  "author": "Douro Labs",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/pyth-network/pyth-crosschain/issues"
+  },
+  "homepage": "https://github.com/pyth-network/pyth-crosschain/tree/main/express_relay/sdk/solidity",
+  "devDependencies": {
+    "prettier": "^2.7.1",
+    "prettier-plugin-solidity": "^1.0.0-rc.1",
+    "abi_generator": "*"
+  }
+}
