+++ target_chains/ethereum/abi_generator/src/generate.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const { generateAbi } = require("./index");
+const contracts = process.argv.slice(2);
+generateAbi(contracts);
